refactor(performance): complete migration to withUserFlowSpan
- Replace all manual wrapper functions with withUserFlowSpan
- Remove obsolete performance transaction modules
- Update all modules: diet, search, recipe, weight, auth
- Add trackUserFlowFunction decorator for future use
- Fix import paths to use ~/shared/config/performance
- Remove unused variables and fix ESLint errors
- All tests passing, code fully migrated to simpler approach
