+++ src/modules/search/application/usecases/cachedSearchCrud.ts
@@ -1,34 +1,19 @@
 import { createCachedSearchRepository } from '~/modules/search/infrastructure/cachedSearchRepository'
-import { trackSearchCache } from '~/shared/performance'
 
 const cachedSearchRepository = createCachedSearchRepository()
 
-export async function isSearchCached(
-  query: string,
-  transactionId?: string | null,
-): Promise<boolean> {
+export async function isSearchCached(query: string): Promise<boolean> {
   const isCached = await cachedSearchRepository.isSearchCached(query)
 
-  trackSearchCache(
-    transactionId ?? null,
-    isCached ? 'hit' : 'miss',
-    `search_${query}`,
-    { query, isCached },
-  )
+  // trackSearchCache removed - using withUserFlowSpan directly now
 
   return isCached
 }
 
-export async function markSearchAsCached(
-  query: string,
-  transactionId?: string | null,
-): Promise<void> {
+export async function markSearchAsCached(query: string): Promise<void> {
   await cachedSearchRepository.markSearchAsCached(query)
 
-  trackSearchCache(transactionId ?? null, 'write', `search_${query}`, {
-    query,
-    operation: 'mark_cached',
-  })
+  // trackSearchCache removed - using withUserFlowSpan directly now
 }
 
 export async function unmarkSearchAsCached(query: string): Promise<void> {
